+++ lib/data/models/recipe_aliment_model.dart
@@ -13,9 +13,9 @@ class RecipeAliment {
   String? dose;
   short? quantity;
   short? calories;
-  float? proteins;
-  float? carbohydrates;
-  float? sugars;
-  float? lipids;
-  float? saturatedFats;
+  double? proteins;
+  double? carbohydrates;
+  double? sugars;
+  double? lipids;
+  double? saturatedFats;
 }
